pyproject.toml
[project]
name = "gif-tool"
version = "0.1.0"
description = "CLI and library to convert videos or a folder of images into GIFs with validations and sensible defaults"
readme = "README.md"
authors = [{ name = "MHM" }]
requires-python = ">=3.10"
dependencies = [
  "opencv-python",
  "imageio",
  "numpy",
  "Pillow",
]

[project.scripts]
gif-tool = "gif_tool.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
